feat(blog): spread post dates, add discoverability from home and footer
- Spread blog post publish dates across May 1-22 so the archive reflects
  a realistic publishing cadence instead of all six posts dated the same day.
- Footer gains a Blog link as the first item in the link row.
- Home page now ends with a "From the blog" teaser section showing the 3
  latest posts with a "View all posts" link. To enable a server-rendered
  teaser alongside the existing client-side badge picker, the home page
  was split: the interactive picker moved to src/components/HomeContent.tsx
  (still a client component) and src/app/page.tsx became a thin server
  shell composing HomeContent and LatestPostsTeaser.
