+++ public/sitemap.xml
@@ -31,38 +31,38 @@
     <priority>0.8</priority>
   </url>
   <url>
-    <loc>https://madebyhuman.iamjarl.com/blog/why-ai-transparency-matters</loc>
-    <lastmod>2026-05-22</lastmod>
-    <changefreq>yearly</changefreq>
-    <priority>0.7</priority>
-  </url>
-  <url>
-    <loc>https://madebyhuman.iamjarl.com/blog/welcome-to-the-blog</loc>
-    <lastmod>2026-05-22</lastmod>
-    <changefreq>yearly</changefreq>
-    <priority>0.7</priority>
-  </url>
-  <url>
-    <loc>https://madebyhuman.iamjarl.com/blog/made-by-human-vs-not-by-ai</loc>
-    <lastmod>2026-05-22</lastmod>
-    <changefreq>yearly</changefreq>
-    <priority>0.7</priority>
-  </url>
-  <url>
-    <loc>https://madebyhuman.iamjarl.com/blog/how-to-signal-human-creativity</loc>
-    <lastmod>2026-05-22</lastmod>
-    <changefreq>yearly</changefreq>
-    <priority>0.7</priority>
-  </url>
-  <url>
     <loc>https://madebyhuman.iamjarl.com/blog/how-to-add-badge-to-nextjs-react</loc>
     <lastmod>2026-05-22</lastmod>
     <changefreq>yearly</changefreq>
     <priority>0.7</priority>
   </url>
   <url>
+    <loc>https://madebyhuman.iamjarl.com/blog/how-to-signal-human-creativity</loc>
+    <lastmod>2026-05-18</lastmod>
+    <changefreq>yearly</changefreq>
+    <priority>0.7</priority>
+  </url>
+  <url>
+    <loc>https://madebyhuman.iamjarl.com/blog/made-by-human-vs-not-by-ai</loc>
+    <lastmod>2026-05-14</lastmod>
+    <changefreq>yearly</changefreq>
+    <priority>0.7</priority>
+  </url>
+  <url>
     <loc>https://madebyhuman.iamjarl.com/blog/art-of-curation-human-touch</loc>
-    <lastmod>2026-05-22</lastmod>
+    <lastmod>2026-05-09</lastmod>
+    <changefreq>yearly</changefreq>
+    <priority>0.7</priority>
+  </url>
+  <url>
+    <loc>https://madebyhuman.iamjarl.com/blog/why-ai-transparency-matters</loc>
+    <lastmod>2026-05-05</lastmod>
+    <changefreq>yearly</changefreq>
+    <priority>0.7</priority>
+  </url>
+  <url>
+    <loc>https://madebyhuman.iamjarl.com/blog/welcome-to-the-blog</loc>
+    <lastmod>2026-05-01</lastmod>
     <changefreq>yearly</changefreq>
     <priority>0.7</priority>
   </url>
