+++ src/content/blog/art-of-curation-human-touch.mdx
@@ -1,7 +1,7 @@
 ---
 title: "The Art of Curation: Keeping the Human Touch in the AI Era"
 description: "Why taste, intent, and critical curation are the most important creative skills in a world of abundant synthetic content."
-date: "2026-05-22"
+date: "2026-05-09"
 author: "Jarl"
 tags: ["philosophy", "creativity"]
 ---
