+++ src/content/blog/how-to-signal-human-creativity.mdx
@@ -1,7 +1,7 @@
 ---
 title: "How to Signal Human Creativity in Your Projects"
 description: "A guide on showcasing human craftsmanship and intention alongside AI tools, and how badges can build trust."
-date: "2026-05-22"
+date: "2026-05-18"
 author: "Jarl"
 tags: ["guide", "creativity"]
 ---
