+++ src/content/blog/made-by-human-vs-not-by-ai.mdx
@@ -1,7 +1,7 @@
 ---
 title: "Made by Human vs. Not By AI: A Positive Approach to Transparency"
 description: "Why we choose to celebrate human creativity and curation instead of defining our work by what it is not."
-date: "2026-05-22"
+date: "2026-05-14"
 author: "Jarl"
 tags: ["philosophy", "comparison"]
 ---
