+++ src/content/blog/welcome-to-the-blog.mdx
@@ -1,7 +1,7 @@
 ---
 title: "Welcome to the Made by Human blog"
 description: "A place to share thinking, updates, and stories from the Made by Human movement."
-date: "2026-05-22"
+date: "2026-05-01"
 author: "Jarl"
 tags: ["announcement"]
 ---
