+++ src/content/blog/why-ai-transparency-matters.mdx
@@ -1,7 +1,7 @@
 ---
 title: "Why Transparency About AI Use Matters"
 description: "In an AI-saturated world, being open about how you create builds trust and honors the human element in work."
-date: "2026-05-22"
+date: "2026-05-05"
 author: "Jarl"
 tags: ["philosophy", "transparency"]
 ---
